Handle missing extra and cleanup
