+++ colabfold/pdb.py
@@ -1,23 +1,3 @@
-from pathlib import Path
-from typing import Union
-
-import numpy as np
-
-
-def set_bfactor(pdb_filename: Union[str, Path], bfac, idx_res, chains):
-    in_file = open(pdb_filename, "r").readlines()
-    out_file = open(pdb_filename, "w")
-    for line in in_file:
-        if line[0:6] == "ATOM  ":
-            seq_id = int(line[22:26].strip()) - 1
-            # FIXME: This is broken somehow but I don't understand how
-            # seq_id = np.where(idx_res == seq_id)[0][0]
-            out_file.write(
-                f"{line[:21]}{chains[seq_id]}{line[22:60]}{bfac[seq_id]:6.2f}{line[66:]}"
-            )
-    out_file.close()
-
-
 def show_pdb(
     use_amber: bool,
     jobname: str,
